Minor changes and add comments for proof
* src/ada/ada_main.adb
Use Unreferenced aspect to indicate which variables are not expected to be
read after being set.
Change assignment to Host_Flags so that it's done to the whole array at once.
Suggestion to model the connection state

* src/ada/socket_interface.adb
Add comments.
Use "Ignore" prefix (there are others) to silence warnings about unused
local variables.

* src/ada/socket_interface.ads
* src/ada/tcp_binding.ads
Use Data'Last >= Data'First instead of Data'Length > 0 in contracts, as the
latter could overflow on a huge Data array, as the index type is a 32 or 64
bits modular type, and Data'Length is implicitly converted to a signed integer
as per SPARK semantics.

* src/ada/udp_binding.ads
Add newline.